refactor: extract common logic
